+++ src/NServiceBus.TransactionalSession.AcceptanceTests/EndpointCustomizationConfigurationExtensions.cs
