--- src/NServiceBus.TransactionalSession.AcceptanceTests/EndpointTemplates/ConfigureExtensions.cs
@@ -1,23 +1,0 @@
-﻿namespace NServiceBus.TransactionalSession.AcceptanceTests.EndpointTemplates
-{
-    using AcceptanceTesting.Support;
-    using Microsoft.Extensions.DependencyInjection;
-
-    public static class ConfigureExtensions
-    {
-        public static void RegisterComponentsAndInheritanceHierarchy(this EndpointConfiguration builder, RunDescriptor runDescriptor)
-        {
-            builder.RegisterComponents(r => { RegisterInheritanceHierarchyOfContextOnContainer(runDescriptor, r); });
-        }
-
-        static void RegisterInheritanceHierarchyOfContextOnContainer(RunDescriptor runDescriptor, IServiceCollection r)
-        {
-            var type = runDescriptor.ScenarioContext.GetType();
-            while (type != typeof(object))
-            {
-                r.AddSingleton(type, runDescriptor.ScenarioContext);
-                type = type.BaseType;
-            }
-        }
-    }
-}
