--- src/NServiceBus.TransactionalSession.AcceptanceTests/EndpointTemplates/DefaultServer.cs
@@ -1,58 +1,0 @@
-﻿namespace NServiceBus.TransactionalSession.AcceptanceTests.EndpointTemplates
-{
-    using System;
-    using System.Collections.Generic;
-    using System.IO;
-    using System.Threading.Tasks;
-    using AcceptanceTesting.Customization;
-    using AcceptanceTesting.Support;
-
-    public class DefaultServer : IEndpointSetupTemplate
-    {
-        public DefaultServer()
-        {
-            typesToInclude = new List<Type>();
-        }
-
-        public DefaultServer(List<Type> typesToInclude)
-        {
-            this.typesToInclude = typesToInclude;
-        }
-
-#pragma warning disable PS0018 // A task-returning method should have a CancellationToken parameter unless it has a parameter implementing ICancellableContext
-#pragma warning disable PS0013 // A Func used as a method parameter with a Task, ValueTask, or ValueTask<T> return type argument should have at least one CancellationToken parameter type argument unless it has a parameter type argument implementing ICancellableContext
-        public async Task<EndpointConfiguration> GetConfiguration(RunDescriptor runDescriptor, EndpointCustomizationConfiguration endpointConfiguration, Func<EndpointConfiguration, Task> configurationBuilderCustomization)
-#pragma warning restore PS0013 // A Func used as a method parameter with a Task, ValueTask, or ValueTask<T> return type argument should have at least one CancellationToken parameter type argument unless it has a parameter type argument implementing ICancellableContext
-#pragma warning restore PS0018 // A task-returning method should have a CancellationToken parameter unless it has a parameter implementing ICancellableContext
-        {
-            var types = endpointConfiguration.GetTypesScopedByTestClass();
-
-            typesToInclude.AddRange(types);
-
-            var configuration = new EndpointConfiguration(endpointConfiguration.EndpointName);
-
-            configuration.TypesToIncludeInScan(typesToInclude);
-            configuration.EnableInstallers();
-
-            configuration.UsePersistence<AcceptanceTestingPersistence>();
-            var storageDir = Path.Combine(NServiceBusAcceptanceTest.StorageRootDir, NUnit.Framework.TestContext.CurrentContext.Test.ID);
-            configuration.UseTransport(new LearningTransport
-            {
-                StorageDirectory = storageDir,
-                TransportTransactionMode = TransportTransactionMode.ReceiveOnly
-            });
-
-            var recoverability = configuration.Recoverability();
-            recoverability.Delayed(delayed => delayed.NumberOfRetries(0));
-            recoverability.Immediate(immediate => immediate.NumberOfRetries(0));
-
-            configuration.RegisterComponentsAndInheritanceHierarchy(runDescriptor);
-
-            await configurationBuilderCustomization(configuration).ConfigureAwait(false);
-
-            return configuration;
-        }
-
-        List<Type> typesToInclude;
-    }
-}
