--- src/NServiceBus.TransactionalSession.AcceptanceTests/NServiceBusAcceptanceTest.cs
@@ -1,75 +1,0 @@
-﻿namespace NServiceBus.TransactionalSession.AcceptanceTests
-{
-    using System;
-    using System.IO;
-    using System.Linq;
-    using System.Threading;
-    using AcceptanceTesting.Customization;
-    using NUnit.Framework;
-
-    /// <summary>
-    /// Base class for all the NSB test that sets up our conventions
-    /// </summary>
-    [TestFixture]
-    public abstract partial class NServiceBusAcceptanceTest
-    {
-        [SetUp]
-        public void SetUp()
-        {
-            Conventions.EndpointNamingConvention = t =>
-            {
-                var classAndEndpoint = t.FullName.Split('.').Last();
-
-                var testName = classAndEndpoint.Split('+').First();
-
-                testName = testName.Replace("When_", "");
-
-                var endpointBuilder = classAndEndpoint.Split('+').Last();
-
-                testName = Thread.CurrentThread.CurrentCulture.TextInfo.ToTitleCase(testName);
-
-                testName = testName.Replace("_", "");
-
-                return testName + "." + endpointBuilder;
-            };
-        }
-
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            if (Directory.Exists(StorageRootDir))
-            {
-                Directory.Delete(StorageRootDir, true);
-            }
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            if (Directory.Exists(StorageRootDir))
-            {
-                Directory.Delete(StorageRootDir, true);
-            }
-        }
-
-        public static string StorageRootDir
-        {
-            get
-            {
-                string tempDir;
-
-                if (Environment.OSVersion.Platform == PlatformID.Win32NT)
-                {
-                    //can't use bin dir since that will be too long on the build agents
-                    tempDir = @"c:\temp";
-                }
-                else
-                {
-                    tempDir = Path.GetTempPath();
-                }
-
-                return Path.Combine(tempDir, "nsb.metrics-att");
-            }
-        }
-    }
-}
