+++ src/NServiceBus.TransactionalSession.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NServiceBus.TransactionalSession", "NServiceBus.TransactionalSession\NServiceBus.TransactionalSession.csproj", "{B4882309-F2E5-4B96-8A98-CFACB30AD117}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus.TransactionalSession.AcceptanceTests", "NServiceBus.TransactionalSession.AcceptanceTests\NServiceBus.TransactionalSession.AcceptanceTests.csproj", "{4994C04C-42CA-467E-922E-AA1E63E8EAE3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NServiceBus.TransactionalSession.AcceptanceTests", "NServiceBus.TransactionalSession.AcceptanceTests\NServiceBus.TransactionalSession.AcceptanceTests.csproj", "{4994C04C-42CA-467E-922E-AA1E63E8EAE3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
